Increased steerRatio(17), Ki,Kp, tire stiffness. corrected weight (S85), front/rear weight balance (48/52). Disabled saturation alert as it does not really pertain to us. Will have to add few alerts for EPAS just in case. Removed cruise control variation related to visual radar.

+++ selfdrive/car/tesla/teslacan.py
@@ -45,7 +45,7 @@ def create_steering_control(enabled, apply_steer, idx):
  type_counter = steering_type << 6
  type_counter += idx
  #change angle to the Tesla * 10 + 0x4000
- apply_steer = ( apply_steer * 10 + 0x4000 ) & 0xFFFF
+ apply_steer = int( apply_steer * 10 + 0x4000 ) & 0xFFFF
  struct.pack_into('!hB', msg, 0,  apply_steer, type_counter)
  struct.pack_into('B', msg, msg_len-1, add_tesla_checksum(msg_id,msg))
  return [msg_id, 0, msg.raw, 2]
